chore: move the src directory to another server crate which will contain bin folder
